The last piece is almost done. Bug fix.

+++ querygo/querygo.go
@@ -22,29 +22,84 @@ func Init() {
 	}
 }
 
-func ExportProject(dir string) error {
+func ExportProject(dir string, dryrun bool, simple bool) error {
 	_gpro, err := golang.ParseProject(dir)
 	if err != nil {
 		return err
 	}
 
-	gpro := gopro(*_gpro)
-	_, err = gpro.Write(DB)
-	if err != nil {
-		if _err := DeleteProject(DB, _gpro.Name); _err != nil {
-			return _err
+	if !dryrun {
+		gpro := gopro(*_gpro)
+
+		if simple {
+			if _err := DeleteProject(DB, _gpro.Name); _err != nil {
+				return _err
+			}
 		}
 
+		_, err = gpro.Write(DB)
+		if err != nil {
+			if _err := DeleteProject(DB, _gpro.Name); _err != nil {
+				return _err
+			}
+
+			return err
+		}
+	}
+
+	return nil
+}
+
+func ExportFile(file *os.File, dryrun bool, simple bool) error {
+	_gfile, err := golang.ParseFile(file)
+	if err != nil {
 		return err
+	}
+
+	if !dryrun {
+		gfile := gofile(*_gfile)
+
+		if simple {
+			if _err := DeleteFile(DB, _gfile.Name); _err != nil {
+				return _err
+			}
+		}
+
+		_, err = gfile.Write(DB)
+		if err != nil {
+			if _err := DeleteFile(DB, _gfile.Name); _err != nil {
+				return _err
+			}
+
+			return err
+		}
+	}
+
+	return nil
+}
+
+func Export(file string, dryrun bool, simple bool) error {
+	if f, err := os.Open(file); err != nil {
+		return err
+	} else {
+		fstat, err := f.Stat()
+		if err != nil {
+			return err
+		}
+		if fstat.IsDir() {
+			return ExportProject(file, dryrun, simple)
+		} else {
+			return ExportFile(f, dryrun, simple)
+		}
 	}
 
 	return nil
 }
 
 func ExportEveryProjectInGoPath() error {
-	return ExportProject(GOPATH)
+	return ExportProject(GOPATH, false, true)
 }
 
 func PrintVersion() {
-
+	fmt.Println("QueryGo Version 1.0")
 }
